Новая модель лайншипа и отображение подобранных мин

+++ Assets/Scripts/MineDash.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class MineDash: Mine {
     
 	public MineDash(): base() {
-		mineType = (int)MineTypes.Dash;
+		mineType = MineTypes.Dash;
         mineTypeColor = Color.red;
     }
 }
