+++ Assets/Scripts/MineSimple.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class MineSimple: Mine {
 
 	public MineSimple(): base() {
-		mineType = (int)MineTypes.Simple;
+		mineType = MineTypes.Simple;
         mineTypeColor = Color.grey;
     }
 }
