+++ Assets/Scripts/MineSpeed.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class MineSpeed: Mine {
 
     public MineSpeed(): base() {
-		mineType = (int)MineTypes.Speed;
+		mineType = MineTypes.Speed;
         mineTypeColor = Color.blue;
     }
 }
